+++ src/pages/cheolho/Main/Main.scss
@@ -117,6 +117,10 @@
     box-sizing: border-box;
     margin-right: 40px;
 
+    .feed-left {
+      margin-bottom: 20px;
+    }
+
     .inner {
       position: relative;
     }
